+++ test/Spec.hs
@@ -23,6 +23,7 @@ import Network.Wai.Handler.Warp qualified as Warp
 import Network.Wreq qualified as Wreq
 import Test.Tasty
 import Test.Tasty.HUnit
+import qualified Keter.Config.V10Spec as V10
 import qualified Keter.Proxy.MiddlewareSpec as ProxyMW
 
 main :: IO ()
@@ -36,6 +37,7 @@ keterTests =
     , testCase "Wildcard Domains" caseWildcards
     , testCase "Head then post doesn't crash" headThenPostNoCrash
     , ProxyMW.tests
+    , V10.tests
     ]
 
 caseSubdomainIntegrity :: IO ()